requirements.txt
requests
pandas
python-dotenv
psycopg2
SQLAlchemy
plotly
dash
gunicorn
waitress
